# [driver] [test] PRNG: allow custom\fixed seed