--- a/driver/conv_driver.hpp
+++ b/driver/conv_driver.hpp
@@ -1075,7 +1075,7 @@ namespace detail {
 template <typename T>
 T RanGenWeights()
 {
-    return RAN_GEN<T>(static_cast<T>(-0.5), static_cast<T>(0.5));
+    return prng::gen_A_to_B(static_cast<T>(-0.5), static_cast<T>(0.5));
 }
 
 // Shift FP16 distribution towards positive numbers,
@@ -1083,7 +1083,14 @@ T RanGenWeights()
 template <>
 float16 RanGenWeights()
 {
-    return RAN_GEN<float16>(static_cast<float16>(-1.0 / 3.0), static_cast<float16>(0.5));
+    return prng::gen_A_to_B(static_cast<float16>(-1.0 / 3.0), static_cast<float16>(0.5));
+}
+
+// int8 has it's own range
+template <>
+int8_t RanGenWeights()
+{
+    return prng::gen_A_to_B(static_cast<int8_t>(-1), static_cast<int8_t>(1));
 }
 
 } // namespace detail
@@ -1304,10 +1311,6 @@ int ConvDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
     std::string biasFileName = inflags.GetValueStr("in_bias");
     std::string doutFileName = inflags.GetValueStr("dout_data");
 
-    /* Unless seed is persistent between runs validation using cache stored in file is impossible.
-     */
-    srand(0);
-
     bool dataRead = false;
     if(is_fwd || is_wrw)
         if(!inFileName.empty())
@@ -1318,36 +1321,17 @@ int ConvDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
         if(!weiFileName.empty())
             weiRead = readBufferFromFile<Tgpu>(wei.data.data(), wei_sz, weiFileName.c_str());
 
+    const Tgpu Data_scale = is_int8 ? static_cast<Tgpu>(127) : static_cast<Tgpu>(0.01);
     if(is_int8)
     {
-        float Data_scale = 127.0;
-
-        if(!dataRead)
-        {
-            for(int i = 0; i < in_sz; i++)
-            {
-                if(is_fwd || is_wrw)
-                    in.data[i] =
-                        static_cast<Tgpu>(Data_scale * RAN_GEN<float>(static_cast<float>(0.0),
-                                                                      static_cast<float>(1.0)));
-                else /// \anchor move_rand
-                    /// Move rand() forward, even if buffer is unused. This provides the same
-                    /// initialization of input buffers regardless of which kinds of
-                    /// convolutions are currently selectedfor testing (see the "-F" option).
-                    /// Verification cache would be broken otherwise.
-                    GET_RAND();
-            }
-        }
-
         if(inflags.GetValueInt("bias") != 0)
         {
             size_t b_sz = GetTensorSize(biasTensor);
             b_dev       = std::unique_ptr<GPUMem>(new GPUMem(ctx, b_sz, sizeof(float)));
-            b_int8      = std::vector<float>(b_sz, static_cast<float>(0));
+            b_int8      = std::vector<float>(b_sz, 0.f);
             for(int i = 0; i < b_sz; i++)
             {
-                b_int8[i] = static_cast<float>(i % 8) +
-                            RAN_GEN<float>(static_cast<float>(0.0), static_cast<float>(1.0));
+                b_int8[i] = static_cast<float>(i % 8) + prng::gen_canonical<float>();
             }
 
             if(!biasFileName.empty())
@@ -1357,46 +1341,22 @@ int ConvDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
 
             b_dev->ToGPU(q, b_int8.data());
         }
-
-        if(!weiRead)
-        {
-            for(int i = 0; i < wei_sz; i++)
-                if(is_fwd || is_bwd)
-                    wei.data[i] =
-                        static_cast<Tgpu>(Data_scale * 2 * detail::RanGenWeights<float>());
-                else /// \ref move_rand
-                    GET_RAND();
-        }
     }
     else
     {
-        Tgpu Data_scale = static_cast<Tgpu>(0.01);
-
         bool doutRead = false;
         if(is_bwd || is_wrw)
             if(!doutFileName.empty())
                 doutRead = readBufferFromFile<Tgpu>(dout.data.data(), out_sz, doutFileName.c_str());
 
-        if(!dataRead)
-        {
-            for(int i = 0; i < in_sz; i++)
-            {
-                if(is_fwd || is_wrw)
-                    in.data[i] =
-                        Data_scale * RAN_GEN<Tgpu>(static_cast<Tgpu>(0.0), static_cast<Tgpu>(1.0));
-                else /// \ref move_rand
-                    GET_RAND();
-            }
-        }
-
         if(!doutRead)
         {
             for(int i = 0; i < out_sz; i++)
+            {
+                auto val = prng::gen_0_to_B(Data_scale);
                 if(is_bwd || is_wrw)
-                    dout.data[i] =
-                        Data_scale * RAN_GEN<Tgpu>(static_cast<Tgpu>(0.0), static_cast<Tgpu>(1.0));
-                else /// \ref move_rand
-                    GET_RAND();
+                    dout.data[i] = val;
+            }
         }
 
         if(inflags.GetValueInt("bias") != 0)
@@ -1409,10 +1369,8 @@ int ConvDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
             db_host     = tensor<Tref>(miopen::deref(biasTensor));
             for(int i = 0; i < b_sz; i++)
             {
-                b.data[i] = static_cast<Tgpu>(i % 8) +
-                            RAN_GEN<Tgpu>(static_cast<Tgpu>(0.0), static_cast<Tgpu>(1.0));
-                db[i] = static_cast<Tgpu>(i % 8) +
-                        RAN_GEN<Tgpu>(static_cast<Tgpu>(0.0), static_cast<Tgpu>(1.0));
+                b.data[i] = static_cast<Tgpu>(i % 8) + prng::gen_canonical<Tgpu>();
+                db[i]     = static_cast<Tgpu>(i % 8) + prng::gen_canonical<Tgpu>();
             }
 
             if(!biasFileName.empty())
@@ -1423,14 +1381,25 @@ int ConvDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
             b_dev->ToGPU(q, b.data.data());
             db_dev->ToGPU(q, db.data());
         }
+    }
 
-        if(!weiRead)
+    if(!dataRead)
+    {
+        for(int i = 0; i < in_sz; i++)
         {
-            for(int i = 0; i < wei_sz; i++)
-                if(is_fwd || is_bwd)
-                    wei.data[i] = Data_scale * detail::RanGenWeights<Tgpu>();
-                else /// \ref move_rand
-                    GET_RAND();
+            auto val = prng::gen_0_to_B(Data_scale);
+            if(is_fwd || is_wrw)
+                in.data[i] = val;
+        }
+    }
+
+    if(!weiRead)
+    {
+        for(int i = 0; i < wei_sz; i++)
+        {
+            auto w = Data_scale * detail::RanGenWeights<Tgpu>();
+            if(is_fwd || is_bwd)
+                wei.data[i] = w;
         }
     }