--- a/driver/ctc_driver.hpp
+++ b/driver/ctc_driver.hpp
@@ -255,7 +255,7 @@ int CTCDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
 
     for(int i = 0; i < labels_sz; i++)
     {
-        labels[i] = static_cast<int>(GET_RAND() % num_class + 1);
+        labels[i] = prng::gen_off_range(1, num_class);
         if(blank_lb > num_class)
             labels[i] = labels[i] == num_class ? num_class - 1 : labels[i];
         else if(blank_lb < 0)
@@ -303,12 +303,11 @@ int CTCDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
     workspace      = std::vector<Tgpu>(workSpaceSize / sizeof(Tgpu), 0);
     workspace_host = std::vector<Tref>(workSpaceSizeCPU / sizeof(Tref), 0);
 
-    srand(0);
     double scale = 0.01;
 
     for(int i = 0; i < probs_sz; i++)
     {
-        probs[i] = static_cast<Tgpu>((static_cast<double>(scale * GET_RAND()) * (1.0 / RAND_MAX)));
+        probs[i] = static_cast<Tgpu>(prng::gen_0_to_B(scale));
     }
     if(apply_softmax)
     {