--- a/driver/gemm_driver.hpp
+++ b/driver/gemm_driver.hpp
@@ -241,7 +241,7 @@ int GemmDriver<T>::AllocateBuffersAndCopy()
 #if GEMM_DRIVER_DEBUG
         a[i] = static_cast<double>(i);
 #else
-        a[i] = static_cast<double>(GET_RAND()) * (1.0 / RAND_MAX);
+        a[i] = prng::gen_canonical<double>();
 #endif
     }
 
@@ -250,7 +250,7 @@ int GemmDriver<T>::AllocateBuffersAndCopy()
 #if GEMM_DRIVER_DEBUG
         b[i] = static_cast<double>(i);
 #else
-        b[i] = static_cast<double>((GET_RAND()) * (1.0 / RAND_MAX) - 0.5) * 0.001;
+        b[i] = prng::gen_A_to_B(-0.5, 0.5) * 0.001;
 #endif
     }
 #if MIOPEN_BACKEND_OPENCL