--- a/driver/lrn_driver.hpp
+++ b/driver/lrn_driver.hpp
@@ -254,7 +254,7 @@ int LRNDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
         {
             for(int i = 0; i < scale.size(); i++)
             {
-                scale[i]     = RAN_GEN<Tgpu>(static_cast<Tgpu>(0.0), static_cast<Tgpu>(1.0));
+                scale[i]     = prng::gen_canonical<Tgpu>();
                 scalehost[i] = Tref(scale[i]);
             }
         }
@@ -265,13 +265,13 @@ int LRNDriver<Tgpu, Tref>::AllocateBuffersAndCopy()
 
     for(int i = 0; i < in_sz; i++)
     {
-        in[i] = RAN_GEN<Tgpu>(static_cast<Tgpu>(-1.0), static_cast<Tgpu>(1.0));
+        in[i] = prng::gen_A_to_B(static_cast<Tgpu>(-1), static_cast<Tgpu>(1));
     }
 
     Tgpu Data_scale = static_cast<Tgpu>(0.001);
     for(int i = 0; i < out_sz; i++)
     {
-        dout[i] = Data_scale * RAN_GEN<Tgpu>(static_cast<Tgpu>(-0.5), static_cast<Tgpu>(0.5));
+        dout[i] = Data_scale * prng::gen_A_to_B(static_cast<Tgpu>(-0.5), static_cast<Tgpu>(0.5));
     }
 
 #if MIOPEN_BACKEND_OPENCL