https://github.com/ROCmSoftwarePlatform/MIOpen/labels/urgency_blocker because we've already got randomly failed cases.
<details>
<summary>For example (GFX1100 required)</summary>

```
MIOPEN_DEBUG_DRIVER_PRNG_SEED=1372743782 ./bin/MIOpenDriver convfp16 -W 1195 -H 16 -c 1 -n 16 -k 1 -x 1 -y 1 --pad_w 12 --pad_h 5 -F 1 -t 1 -i 1 -V 1 -S 139
```

</details>

Significant PRNG re-design.

Highlights:
- switch `std::linear_congruential_engine` parameters to glibc version (speedup)
- simplify canonical number generation (speedup)
- simplify uniform distribution generation (speedup)
- provide new primitives to generate random numbers (code quality, speedup)
- remove lots of copy-pasted generators and replaced them with new primitives (code quality)
- "hide" direct access to PRNG (code quality)
- fix verification cache, at least partially
- drop legacy generator support (code quality)
- unify test and driver generators (code quality) - I see no reasons to split them:
initially they were dependent through global `rand()` state and through including files from `test` to `driver`
most probably they will suffer from the same problems, so will be updated simultaneously and I can hardly imagine separated updates. But I'm not quite sure about this unification, probably I've missed something.

I did some simple performance measurements for `AllocateBuffersAndCopy` for convolution and got following results (compare to the current develop):

`AllocateBuffersAndCopy`:
|case|old (ms)|new (ms)|gain|
|-|-|-|-|
|-N 128|1253.08|936.23|1.34|
|-N 1|43.57|47.39|1.09|

Depending on the case allocation may take from 60% to 10% of total execution time and thus total improvement may not be so significant.

Total time:
|case|old (ms)|new (ms)|gain|
|-|-|-|-|
|-N 128 -V 1|3056|3328|1.09|
|-N 1 -V 1|1709|1726|1.01|
|-N 128 -V 0|1423|1756|1.23|
|-N 1 -V 0|464|476|1.03|

<details>
<summary>The other driver parameters</summary>

```
./bin/MIOpenDriver conv --spatial_dim 2 --pad_h 0 --pad_w 0 --pad_d 0 --conv_stride_h 1 --conv_stride_w 1 --conv_stride_d 0 --dilation_h 1 --dilation_w 1 --dilation_d 0 --group_count 1 --mode conv --pad_mode default --trans_output_pad_h 0 --trans_output_pad_w 0 --trans_output_pad_d 0 --out_layout NCHW --in_d 1 --in_h 210 --in_w 210 --fil_d 1 --fil_h 3 --fil_w 3 --in_channels 16 --out_channels 16 --forw 1 -i 10 -t 1 -w 1 -s 84
```

</details>

I haven't tried it on the other cases or algorithms, but the overall trend seems to be: bigger tensors and no verification (perf runs) - more speedup.

@atamazov please join.

Closes #2313