@JehandadKhan 2023-08-25
@CAHEK7  Test `test_rnn_vanilla_dropout ` fails FP16 on MI100.