@CAHEK7 2023-08-25
Recently for some reasons the tests often failed on a git-checkout stage.