@junliume 2023-08-31
This list of tests is becoming unstable with this update on major platforms:
test_conv_igemm_dynamic_xdlops_nhwc_wrw_nchw
test_conv2d
test_conv_for_implicit_gemm
test_immed_conv2d